タスクコンポーネントを分割
