+++ src/components/Task.jsx
@@ -1,30 +1,15 @@
 import { useState } from "react";
-import { TaskList,TaskInput,TaskDone,TaskDoneText,DeleteIcon } from "../css/task";
+import { TaskList,TaskInput} from "../css/task";
+import { Delete } from "./task-components/TaskDelete";
+import { TaskComplete } from "./task-components/TaskDone";
+
 
 export const Task = (props) => {
   const { taskText,taskList,id,setTaskList,showList} = props;
-  
   const [taskDone, setTaskDone] = useState(false);
 
   const onChangeTaskInput = () => {
     console.log('change task input');
-  }
-
-  const onClickDelete = () => {
-    const deleteTask = [...taskList];
-    
-    const delateOK = window.confirm('todoを削除してもいいですか？');
-    if(delateOK === true) {
-      deleteTask.splice(id,1);
-      setTaskList(deleteTask);
-    }
-  }
-
-  const onClickDone = () => {
-    setTaskDone(true);
-    if(taskDone === true) {
-      setTaskDone(false)
-    }
   }
 
   return (
@@ -39,19 +24,17 @@ export const Task = (props) => {
       style={taskDone ? {backgroundColor:"#BDBDBD",color:'#787878'}: {backgroundColor:"#FFFF"}} 
       />
 
-     <TaskDone style={taskDone ? {backgroundColor:'#BDBDBD'}: {}}>
-       <TaskDoneText 
-        className="sm:text-lg text-sm" 
-        onClick={() => onClickDone()} 
-        style={taskDone ? {color:'#787878'} :{color:'#FFFF'}}
-        >
-        {taskDone ? 'BACK' :'DONE'}
-        </TaskDoneText>
-     </TaskDone>
 
-     <DeleteIcon className="sm:text-xl text-lg sm:ml-3 ml-6" onClick={onClickDelete}>
-     <i className="sm:text-2xl text-xl fas fa-backspace"></i>
-     </DeleteIcon>
+     <TaskComplete
+      taskDone={taskDone}
+      setTaskDone={setTaskDone}
+     />
+
+     <Delete
+      taskList={taskList}
+      id={id}
+      setTaskList={setTaskList}
+      />
 
     </TaskList>
   </div>
@@ -59,6 +42,3 @@ export const Task = (props) => {
    </>
   )
 }
-
-// ,taskDone === true ? {backgroundColor:"#AA93DB"}: {backgroundColor:"#FFFF"}
-// style={taskDone === true ? {backgroundColor:"#AA93DB"}: {backgroundColor:"#FFFF"}}
