+++ src/components/task-components/TaskDone.jsx
@@ -1,0 +1,29 @@
+
+import {TaskDone,TaskDoneText} from "../../css/task";
+
+
+// タスクを完了を押した時の挙動
+export const TaskComplete = (props) => {
+
+  const { taskDone,setTaskDone} = props;
+
+  const onClickDone = () => {
+    setTaskDone(true);
+    if(taskDone === true) {
+      setTaskDone(false)
+    }
+  }
+
+  return(
+    <TaskDone style={taskDone ? {backgroundColor:'#BDBDBD'}: {}}>
+       <TaskDoneText 
+        className="sm:text-lg text-sm" 
+        onClick={() => onClickDone()} 
+        style={taskDone ? {color:'#787878'} :{color:'#FFFF'}}
+        >
+        {taskDone ? 'BACK' :'DONE'}
+        </TaskDoneText>
+     </TaskDone>
+  )
+}
+
